+++ public_html/packages/metafox/core/src/Listeners/UserDeletedListener.php
@@ -7,13 +7,20 @@ use MetaFox\Platform\Contracts\User;
 
 class UserDeletedListener
 {
-    public function handle(User $user): void
+    public function handle(?User $user): void
     {
+        if (!$user) {
+            return;
+        }
+
         $this->deleteLinks($user);
     }
 
-    protected function deleteLinks(User $user): void
+    protected function deleteLinks(?User $user): void
     {
+        if (!$user) {
+            return;
+        }
         $repository = resolve(LinkRepositoryInterface::class);
 
         $repository->deleteUserData($user->entityId());
