+++ public_html/packages/metafox/marketplace/src/Listeners/UserDeletedListener.php
@@ -8,12 +8,18 @@ use MetaFox\Platform\Contracts\User;
 
 class UserDeletedListener
 {
-    public function handle(User $user): void
+    public function handle(?User $user): void
     {
+        if (!$user) {
+            return;
+        }
         $listingService = resolve(ListingRepositoryInterface::class);
         $listingService->deleteUserData($user);
         $listingService->deleteOwnerData($user);
 
-        resolve(ListingHistoryRepositoryInterface::class)->deleteHistoriesByUser($user->entityId(), $user->entityType());
+        resolve(ListingHistoryRepositoryInterface::class)->deleteHistoriesByUser(
+            $user->entityId(),
+            $user->entityType()
+        );
     }
 }
